+++ Engine/src/UnitActions/Move/MoveNorthWest.cs
@@ -1,21 +1,20 @@
 namespace Civ2engine.UnitActions.Move
 {
-    public class MoveSouthWest : IGameAction
+    public class MoveNorthWest : IGameAction
     {
         public void TriggerAction()
         {
-            if (Game.Instance.CurrentMap.ActiveXY[1] >= Game.Instance.CurrentMap.YDim -1)
+            if (Game.Instance.CurrentMap.ActiveXY[1] == 0)
             {
                 return;
             }
-            
             if (Game.Instance.CurrentMap.ActiveXY[0] > 0)
             {
-                MovementFunctions.MoveC2(-1, 1);
+                MovementFunctions.MoveC2(-1, -1);
             }
             else if(!Game.Instance.Options.FlatEarth)
             {
-                MovementFunctions.MoveC2(Game.Instance.CurrentMap.XDim*2-2, 1);
+                MovementFunctions.MoveC2(Game.Instance.CurrentMap.XDim*2-1, -1);
             }
         }
     }
